% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkmate}
\alias{checkmate}
\title{Checkmate}
\usage{
checkmate(game, color = "W")
}
\arguments{
\item{game}{information about the game}

\item{color}{a string argument that takes a letter of 'W' for White
and 'B' for black.}
}
\value{
A boolean logic of TRUE or FALSE
}
\description{
This function helps to find the games that ended in a checkmate
}
