% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{games}
\alias{games}
\title{games}
\format{
An object of class \code{data.frame} with 20058 rows and 11 columns.
}
\usage{
games
}
\description{
A dataset of the lichess games from kaggle.
}
\keyword{datasets}
