% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{games_clean}
\alias{games_clean}
\title{games_clean}
\format{
An object of class \code{list} of length 19619.
}
\usage{
games_clean
}
\description{
A dataset of cleaned chess games.
}
\keyword{datasets}
