% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initialize_board}
\alias{initialize_board}
\title{Helper function to store a board when the king is checkmated}
\usage{
initialize_board()
}
\value{
A list with the board and board indices
}
\description{
This function initializes a board and stores information about what locations
  pieces are killed at.
}
