% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{king_pos}
\alias{king_pos}
\title{King Position}
\usage{
king_pos(game)
}
\arguments{
\item{game}{information about the game}
}
\value{
A boolean logic of TRUE or FALSE
}
\description{
This function helps to find the kings position
}
