% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(
  split_games,
  lichess_games,
  num_games = length(split_games),
  estimator = 100
)
}
\arguments{
\item{split_games}{The lichess dataset moves for each game split by spaces}

\item{lichess_games}{the lichess dataset}

\item{num_games}{the number of games to be processed}

\item{estimator}{the number of iterations used to predict the time until
completion.}
}
\value{
A list of games where each game is a list that includes the pgn,
  the history, and other traits for subsetting.
}
\description{
Takes the data from the lichess dataset on kaggle and converts
  them into rchess objects and then extracts desired information in a
  memory efficient manner
}
\examples{
print("FIXME LATER")
}
