% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{piece_death}
\alias{piece_death}
\title{Helper function of}
\usage{
piece_death(game, piece = "r", color = "w", position = "l")
}
\arguments{
\item{game}{information about the game}

\item{color}{a string argument that takes a letter of 'W' for White
and 'B' for black.}
}
\value{
A boolean logic of TRUE or FALSE
}
\description{
This function helps to find the games that involved a checkmate of the
  and a time delay between moves and generates the sequence of chess moves
  that make up that specific game. Above the game board a scoreboard is
  provided to track the number points each player has as the game progresses.
}
