% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{piece_pos}
\alias{piece_pos}
\title{Helper function to find death of a piece in a chess game}
\usage{
piece_pos(game, piece = "r", color = "w", position = "l")
}
\arguments{
\item{game}{information about the game}

\item{piece}{a lower case character input of one of the following:
'r' for rook, 'n' for knight, 'b' for bishop.}

\item{color}{a string argument that takes a letter of 'W' for White
and 'B' for black.}

\item{position}{a character of 'l' for left and 'r' for right side}
}
\value{
A boolean logic of TRUE or FALSE
}
\description{
This function takes the users input of game, piece, color, and position and
  finds where that piece died in a game.
}
