% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_game_score.R
\name{plot_game_score}
\alias{plot_game_score}
\title{Plot Chess Game with Score}
\usage{
plot_game_score(game_num = 1, delay = 0.5)
}
\arguments{
\item{game_num}{The row number of the game dataset for plotting.}

\item{delay}{The time in seconds between each plot of a new chess move.}
}
\value{
A sequence of chess moves characterizing the desired game, with a
  scoreboard displayed above it.
}
\description{
Takes a chess game number associated with the games dataset
  and a time delay between moves and generates the sequence of chess moves
  that make up that specific game. Above the game board a scoreboard is
  provided to track the number points each player has as the game progresses.
}
\examples{
plot_game_score(game_num = 1, delay = 0)
}
