% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score.R
\name{plot_score}
\alias{plot_score}
\title{Plot Chess Score}
\usage{
plot_score(game_num = 1)
}
\arguments{
\item{game_num}{The row number of the game dataset for scoring.}
}
\value{
A plot of chess scores across moves for the desired game.
}
\description{
Takes a chess game number associated with the games dataset
  and generates a plot of the score differential across moves.
}
\examples{
plot_score(game_num = 100)
}
